// frontend/apps/frontend/src/app/swagger/model/seatDto.ts
/**
 * CorePlugin.BackendDevServer
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SeatDto {
    nr: number;
    row: number;
    col: number;
    isToFront: boolean;
    isBlocked: boolean;
}